CLN: refactor test utils (#13)
